authorizer/custom/requirements.txt
pyjwt==2.9.0
